+++ examples/data-sources/host_group/data-source.tf
@@ -1,8 +1,0 @@
-data "host_group" "admin" {
-  role = "admin"
-}
-
-resource "host_user" "workstation" {
-  username = "workstation"
-  groups   = [data.host_group.admin.name]
-}