+++ internal/provider/provider_data.go
@@ -7,7 +7,6 @@ type HostProviderData struct {
 	PackageManager        PackageManager
 	BrewManager           BrewPackageManager
 	ScheduleManager       ScheduleManager
-	IdentityManager       IdentityManager
 	GitPath               string
 	SSHKeyManager         SSHKeyManager
 	MacOSDefaultsManager  MacOSDefaultsManager